Move the class Time to the package utility
